Remove intermediary ItemDisplay comp and rename ItemList to SubList
